# services/gateway/public/swagger/openapi.yaml
openapi: 3.0.0
info:
  title: 'eSapiens Challenge API'
  contact:
    name: 'Guilherme Alves'
    email: guihalves20@gmail.com
  version: '1.0'
paths:
...
  '/comments/post/{postID}':
    get:
      tags:
        - Comments
      summary: 'Lista todas as publicações de um postagem'
      operationId: 'App\Http\Controllers\CommentController::indexByPost'
      parameters:
        -
          name: email
          in: query
          description: 'E-mail do usuário - Usado como login'
          required: true
          schema:
            type: string
        -
          name: password
          in: query
          description: 'Senha do usuário'
          required: true
          schema:
            type: string
        -
          name: postID
          in: path
          description: 'ID do post'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
    delete:
      tags:
        - Comments
      summary: 'Deleta TODOS os comentários de uma postagem'
      operationId: 'App\Http\Controllers\CommentController::deleteByPost'
      parameters:
        -
          name: email
          in: query
          description: 'E-mail do usuário - Usado como login'
          required: true
          schema:
            type: string
        -
          name: password
          in: query
          description: 'Senha do usuário'
          required: true
          schema:
            type: string
        -
          name: postID
          in: path
          description: 'ID do post'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
        '403':
          description: 'Autorização negada'
        '500':
          description: 'Erro ao deletar'
  '/comments/user/{userID}':
    get:
      tags:
        - Comments
      summary: 'Lista todas as publicações de um usuário'
      operationId: 'App\Http\Controllers\CommentController::indexByUser'
      parameters:
        -
          name: email
          in: query
          description: 'E-mail do usuário - Usado como login'
          required: true
          schema:
            type: string
        -
          name: password
          in: query
          description: 'Senha do usuário'
          required: true
          schema:
            type: string
        -
          name: userID
          in: path
          description: 'ID do user'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
  '/comments/{postID}/{userID}':
    delete:
      tags:
        - Comments
      summary: 'Deleta TODOS os comentários de um usuário em uma postagem'
      operationId: 'App\Http\Controllers\CommentController::deleteByPostAndUser'
      parameters:
        -
          name: email
          in: query
          description: 'E-mail do usuário - Usado como login'
          required: true
          schema:
            type: string
        -
          name: password
          in: query
          description: 'Senha do usuário'
          required: true
          schema:
            type: string
        -
          name: postID
          in: path
          description: 'ID do post'
          required: true
          schema:
            type: integer
        -
          name: userID
          in: path
          description: 'ID do user'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
        '403':
          description: 'Autorização negada'
        '500':
          description: 'Erro ao deletar'
  '/comments/{commentID}':
    delete:
      tags:
        - Comments
      summary: 'Deleta um comentário'
      parameters:
        -
          name: email
          in: query
          description: 'E-mail do usuário - Usado como login'
          required: true
          schema:
            type: string
        -
          name: password
          in: query
          description: 'Senha do usuário'
          required: true
          schema:
            type: string
        -
          name: commentID
          in: path
          description: 'ID do comentário'
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
        '403':
          description: 'Autorização negada'
        '500':
          description: 'Erro ao deletar'
  /comments:
    post:
      tags:
        - Comments
      summary: 'Cria um comentário em uma postagem'
      parameters:
        -
          name: email
          in: query
          description: 'E-mail do usuário - Usado como login'
          required: true
          schema:
            type: string
        -
          name: password
          in: query
          description: 'Senha do usuário'
          required: true
          schema:
            type: string
        -
          name: post_id
          in: query
          description: 'ID do post'
          required: true
          schema:
            type: integer
        -
          name: coins
          in: query
          description: 'Coins usadas para dar destaque ao comentário'
          required: true
          schema:
            type: integer
        -
          name: title
          in: query
          description: 'Título do comentário'
          required: true
          schema:
            type: string
        -
          name: content
          in: query
          description: 'Conteúdo do comentário'
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
        '400':
          description: 'Dados inválidos'
        '403':
          description: 'Autorização negada'
        '500':
          description: 'Erro ao criar'
...